Add constants for Type and Format. Remove all the wrapper methods for repository formats
