+++ client.go
@@ -20,17 +20,6 @@ type Client interface {
 	ContentType() string
 	ContentTypeTextPlain()
 	ContentTypeJSON()
-	RepositoryAptCreate(Repository, string) error
-	RepositoryAptUpdate(string, Repository, string) error
-	RepositoryAptDelete(string) error
-	RepositoryBowerCreate(Repository, string) error
-	RepositoryBowerRead(string, string) (*Repository, error)
-	RepositoryBowerUpdate(string, Repository, string) error
-	RepositoryBowerDelete(string) error
-	RepositoryDockerCreate(Repository, string) error
-	RepositoryDockerRead(string, string) (*Repository, error)
-	RepositoryDockerUpdate(string, Repository, string) error
-	RepositoryDockerDelete(string) error
 	RepositoryCreate(Repository, string, string) error
 	RepositoryUpdate(string, Repository, string, string) error
 	RepositoryRead(string) (*Repository, error)
