--- repository_apt_test.go
@@ -1,46 +1,0 @@
-package client
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func getTestRepositoryApt(name string) Repository {
-	return Repository{
-		RepositoryApt: &RepositoryApt{
-			Distribution: "bionic",
-		},
-		RepositoryAptSigning: &RepositoryAptSigning{
-			Keypair:    "string",
-			Passphrase: "string",
-		},
-		Name:   name,
-		Online: true,
-		RepositoryCleanup: &RepositoryCleanup{
-			PolicyNames: []string{"weekly-cleanup"},
-		},
-		RepositoryStorage: &RepositoryStorage{
-			BlobStoreName:               "default",
-			StrictContentTypeValidation: true,
-			WritePolicy:                 "allow_once",
-		},
-	}
-}
-
-func TestRepositoryAptHosted(t *testing.T) {
-	client := NewClient(getDefaultConfig())
-	repo := getTestRepositoryApt("test-apt-repo-hosted")
-
-	err := client.RepositoryAptCreate(repo, "hosted")
-	assert.Nil(t, err)
-
-	updatedRepo := repo
-	updatedRepo.Online = false
-
-	err = client.RepositoryAptUpdate(repo.Name, updatedRepo, "hosted")
-	assert.Nil(t, err)
-
-	err = client.RepositoryAptDelete(repo.Name)
-	assert.Nil(t, err)
-}
